Improve indexing of the database (#413)
* Added a option to load the hole database

* Fixed some unit tests

* Implement LoadDatabaseStructure in PxFileDataSource

Replaced the NotImplementedException in the LoadDatabaseStructure method with logic to load the menu from "Menu.xml". The method now constructs the file path, initializes an XmlMenu instance, and returns the CurrentItem based on the specified language.

* Removed some todos

* Refactor Indexer and simplify PxFileDataSource

Updated `Indexer.cs` to use pattern matching for item casting, enhancing readability. Changed `GetCategory` to static. In `PxFileDataSource.cs`, removed unnecessary field assignment in the constructor, streamlining dependencies.

* Refactor LuceneIndex to ensure document updates execute

Removed early return in LuceneIndex.cs, allowing the document update logic to run consistently when the writer is not null. This change improves the handling of document updates in the indexing process.

* Update API test to expect Accepted status code

Modified the assertion in `TableApiControllerTest.cs` to check for `HttpStatusCode.Accepted` (202) instead of `HttpStatusCode.OK` (200). This change reflects an updated expected outcome for the API call to the `/tables?lang=en` endpoint, indicating a potential change in the API's behavior.

* Update HTTP status code expectation in tests

Modified the `TableApiControllerTest.cs` to change the expected HTTP status code for the `/tables?lang=en` endpoint from `HttpStatusCode.Accepted` (202) to `HttpStatusCode.OK` (200), reflecting the correct expected response for successful requests.

* small patch

---------

Co-authored-by: Johannes <jfi@ssb.no>